[CodecIMX] move common IMX buffer definition to imx's global defs - IMX.h
